renamed files and deleted files
