--- app/config/connection.js
@@ -1,19 +1,0 @@
-var mysql = require("mysql");
-
-var connection = mysql.createConnection({
-    host:"localhost",
-    port:3306,
-    user:"root",
-    password:"",
-    database:"caffeinParadise_db"
-});
-
-connection.connect(function(err){
-    if (err){
-        console.log("error connecting " + err.stack);
-        return;
-    }
-    console.log("connected as id " + connection.threadId);
-});
-
-module.exports = connection;
